src/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
playwright>=1.40.0
openai>=1.0.0
azure-ai-inference>=1.0.0b1
python-dotenv>=1.0.0
easyocr>=1.7.0
PyMuPDF>=1.23.0
pillow>=10.0.0
pypdf>=3.0.0
python-docx>=1.1.0